+++ common/commands/src/validator/vesting/create_vesting_schedule.rs
@@ -38,7 +38,7 @@ pub struct Args {
 
     #[clap(
         long,
-        help = "Specify pledge cap as either absolute uNYM value or percentage, flgitoats need to be in the 0.0 to 1.0 range and will be parsed as percentages, integers will be parsed as uNYM"
+        help = "Pledge cap as either absolute uNYM value or percentage, floats need to be in the 0.0 to 1.0 range and will be parsed as percentages, integers will be parsed as uNYM"
     )]
     pub pledge_cap: Option<PledgeCap>,
 }