+++ contracts/vesting/src/vesting/account/delegating_account.rs
@@ -38,7 +38,7 @@ impl DelegatingAccount for Account {
     ) -> Result<Response, ContractError> {
         let current_balance = self.load_balance(storage)?;
         let total_pledged_after = self.total_pledged_locked(storage, env)? + coin.amount;
-        let locked_pledge_cap = self.absolute_pledge_cap(storage)?;
+        let locked_pledge_cap = self.absolute_pledge_cap()?;
 
         if locked_pledge_cap < total_pledged_after {
             return Err(ContractError::LockedPledgeCapReached {